+++ client/src/config.rs
@@ -84,7 +84,6 @@ impl From<&RuntimeConfig> for MaintenanceConfig {
 			replication_factor: val.libp2p.kademlia.record_replication_factor.get() as u16,
 			query_timeout: val.libp2p.kademlia.query_timeout,
 			pruning_interval: val.libp2p.kademlia.store_pruning_interval,
-			telemetry_flush_interval: val.otel.ot_flush_block_interval,
 			automatic_server_mode: val.libp2p.kademlia.automatic_server_mode,
 			total_memory_gb_threshold: val.total_memory_gb_threshold,
 			num_cpus_threshold: val.num_cpus_threshold,