+++ fat/src/main.rs
@@ -8,7 +8,9 @@ use avail_light_core::{
 		Network,
 	},
 	shutdown::Controller,
-	telemetry::{self, otlp::Metrics, MetricCounter, MetricValue},
+	telemetry::{
+		self, otlp::Metrics, MetricCounter, MetricValue, ATTRIBUTE_MULTIADDRESS, ATTRIBUTE_RPC_HOST,
+	},
 	types::{BlockVerified, ClientChannels, IdentityConfig, Origin, ProjectName},
 	utils::{default_subscriber, install_panic_hooks, json_subscriber, spawn_in_span},
 };
@@ -158,7 +160,6 @@ async fn run(config: Config, db: DB, shutdown: Controller<String>) -> Result<()>
 
 	let (maintenance_sender, maintenance_receiver) = mpsc::unbounded_channel::<MaintenanceEvent>();
 	spawn_in_span(shutdown.with_cancel(maintenance::run(
-		config.otel.ot_flush_block_interval,
 		block_rx,
 		shutdown.clone(),
 		maintenance_sender,
@@ -195,29 +196,32 @@ async fn run(config: Config, db: DB, shutdown: Controller<String>) -> Result<()>
 		shutdown.clone(),
 	)));
 
+	let rpc_host = db
+		.get(RpcNodeKey)
+		.map(|node| node.host)
+		.ok_or_else(|| eyre!("No connected host found"))?;
+
 	let metric_attributes = vec![
-		("role".to_string(), "fat".to_string()),
-		("origin".to_string(), Origin::FatClient.to_string()),
-		("version".to_string(), version.to_string()),
-		("peerID".to_string(), p2p_peer_id.to_string()),
-		("partition_size".to_string(), partition_size),
-		("network".to_string(), Network::name(&config.genesis_hash)),
-		("operating_mode".to_string(), "client".to_string()),
+		("role", "fat".to_string()),
+		("origin", Origin::FatClient.to_string()),
+		("version", version.to_string()),
+		("peerID", p2p_peer_id.to_string()),
+		("partition_size", partition_size),
+		("network", Network::name(&config.genesis_hash)),
+		("operating_mode", "client".to_string()),
+		(ATTRIBUTE_MULTIADDRESS, String::default()),
+		(ATTRIBUTE_RPC_HOST, rpc_host),
 	];
 
 	let metrics = telemetry::otlp::initialize(
 		ProjectName::new("avail".to_string()),
 		&Origin::FatClient,
 		config.otel.clone(),
+		metric_attributes,
 	)
 	.wrap_err("Unable to initialize OpenTelemetry service")?;
 
-	let rpc_host = db
-		.get(RpcNodeKey)
-		.map(|node| node.host)
-		.ok_or_else(|| eyre!("No connected host found"))?;
-
-	let mut state = FatState::new(metrics, String::default(), rpc_host, metric_attributes);
+	let mut state = FatState::new(metrics);
 
 	spawn_in_span(shutdown.with_cancel(async move {
 		state
@@ -280,44 +284,15 @@ impl BlockStat {
 
 struct FatState {
 	metrics: Metrics,
-	multiaddress: String,
-	rpc_host: String,
-	metric_attributes: Vec<(String, String)>,
 	active_blocks: HashMap<u32, BlockStat>,
 }
 
 impl FatState {
-	fn new(
-		metrics: Metrics,
-		multiaddress: String,
-		rpc_host: String,
-		metric_attributes: Vec<(String, String)>,
-	) -> Self {
+	fn new(metrics: Metrics) -> Self {
 		FatState {
 			metrics,
-			multiaddress,
-			rpc_host,
-			metric_attributes,
 			active_blocks: Default::default(),
 		}
-	}
-
-	fn update_multiaddress(&mut self, value: String) {
-		self.multiaddress = value;
-	}
-
-	fn update_rpc_host(&mut self, value: String) {
-		self.rpc_host = value;
-	}
-
-	fn attributes(&self) -> Vec<(String, String)> {
-		let mut attrs = vec![
-			("multiaddress".to_string(), self.multiaddress.clone()),
-			("rpc_host".to_string(), self.rpc_host.to_string()),
-		];
-
-		attrs.extend(self.metric_attributes.clone());
-		attrs
 	}
 
 	fn get_block_stat(&mut self, block_num: u32) -> Result<&mut BlockStat> {
@@ -402,38 +377,37 @@ impl FatState {
 		mut fat_receiver: UnboundedReceiver<FatEvent>,
 		mut rpc_receiver: broadcast::Receiver<RpcEvent>,
 	) {
-		self.metrics.count(MetricCounter::Starts, self.attributes());
+		self.metrics.count(MetricCounter::Starts);
 		loop {
 			select! {
 				Some(p2p_event) = p2p_receiver.recv() => {
 					match p2p_event {
 						P2pEvent::Count => {
-							self.metrics.count(MetricCounter::EventLoopEvent, self.attributes());
+							self.metrics.count(MetricCounter::EventLoopEvent)
 						},
 						P2pEvent::IncomingGetRecord => {
-							self.metrics.count(MetricCounter::IncomingGetRecord, self.attributes());
+							self.metrics.count(MetricCounter::IncomingGetRecord)
 						},
 						P2pEvent::IncomingPutRecord => {
-							self.metrics.count(MetricCounter::IncomingPutRecord, self.attributes());
+							self.metrics.count(MetricCounter::IncomingPutRecord)
 						},
 						P2pEvent::Ping(rtt) => {
-							self.metrics.record(MetricValue::DHTPingLatency(rtt.as_millis() as f64))
-								;
+							self.metrics.record(MetricValue::DHTPingLatency(rtt.as_millis() as f64));
 						},
 						P2pEvent::IncomingConnection => {
-							self.metrics.count(MetricCounter::IncomingConnections, self.attributes());
+							self.metrics.count(MetricCounter::IncomingConnections)
 						},
 						P2pEvent::IncomingConnectionError => {
-							self.metrics.count(MetricCounter::IncomingConnectionErrors, self.attributes());
+							self.metrics.count(MetricCounter::IncomingConnectionErrors)
 						},
 						P2pEvent::MultiaddressUpdate(address) => {
-							self.update_multiaddress(address.to_string());
+							self.metrics.set_attribute(ATTRIBUTE_MULTIADDRESS, address.to_string())
 						},
 						P2pEvent::EstablishedConnection => {
-							self.metrics.count(MetricCounter::EstablishedConnections, self.attributes());
+							self.metrics.count(MetricCounter::EstablishedConnections)
 						},
 						P2pEvent::OutgoingConnectionError => {
-							self.metrics.count(MetricCounter::OutgoingConnectionErrors, self.attributes());
+							self.metrics.count(MetricCounter::OutgoingConnectionErrors)
 						},
 						P2pEvent::PutRecord { block_num, records } => {
 							self.handle_new_put_record(block_num, records);
@@ -460,25 +434,15 @@ impl FatState {
 				}
 				Some(maintenance_event) = maintenance_receiver.recv() => {
 					match maintenance_event {
-						MaintenanceEvent::FlushMetrics(block_num) => {
-							if let Err(error) = self.metrics.flush(self.attributes()) {
-								error!(
-									block_num,
-									"Could not handle Flush Maintenance event properly: {error}"
-								);
-							} else {
-								info!(block_num, "Flushing metrics finished");
-							};
-						},
 						MaintenanceEvent::CountUps => {
-							self.metrics.count(MetricCounter::Up, self.attributes());
+							self.metrics.count(MetricCounter::Up);
 						},
 					}
 				}
 				Some(fat_event) = fat_receiver.recv() => {
 					match fat_event {
 						FatEvent::CountSessionBlocks => {
-							self.metrics.count(MetricCounter::SessionBlocks, self.attributes());
+							self.metrics.count(MetricCounter::SessionBlocks);
 						},
 						FatEvent::RecordBlockHeight(block_num) => {
 							self.metrics.record(MetricValue::BlockHeight(block_num));
@@ -494,7 +458,7 @@ impl FatState {
 
 				Ok(rpc_event) = rpc_receiver.recv() => {
 					if let RpcEvent::ConnectedHost(host) = rpc_event {
-						self.update_rpc_host(host);
+						self.metrics.set_attribute(ATTRIBUTE_RPC_HOST, host);
 					}
 				}
 				// break the loop if all channels are closed
@@ -511,12 +475,10 @@ mod maintenance {
 	use tracing::{error, info};
 
 	pub enum OutputEvent {
-		FlushMetrics(u32),
 		CountUps,
 	}
 
 	pub async fn run(
-		ot_flush_block_interval: u32,
 		mut block_receiver: broadcast::Receiver<BlockVerified>,
 		shutdown: Controller<String>,
 		event_sender: UnboundedSender<OutputEvent>,
@@ -525,19 +487,7 @@ mod maintenance {
 
 		loop {
 			match block_receiver.recv().await.map_err(Report::from) {
-				Ok(block) => {
-					let block_num = block.block_num;
-					if block_num % ot_flush_block_interval == 0 {
-						info!(block_num, "Flushing metrics...");
-						if let Err(error) = event_sender.send(OutputEvent::FlushMetrics(block_num))
-						{
-							let error_msg =
-								format!("Failed to send FlushMetrics event: {:#}", error);
-							error!("{error_msg}");
-							_ = shutdown.trigger_shutdown(error_msg);
-							break;
-						}
-					};
+				Ok(_) => {
 					if let Err(error) = event_sender.send(OutputEvent::CountUps) {
 						let error_msg = format!("Failed to send CountUps event: {:#}", error);
 						error!("{error_msg}");