+++ client/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.12.5
+
+- Update `avail-light-core` to 1.2.0
+- Remove `ot_flush_block_interval` from configuration
+
 ## [1.12.4](https://github.com/availproject/avail-light/releases/tag/avail-light-client-v1.12.4) - 2024-12-20
 
 - Update `avail-light-core` to 1.1.0